+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Hitting a mob more than 40 blocks away with an arrow now plays the ding sound
 * Timber Trees now breaks some leaves before falling (35% for Quark's Trumpet trees, 10% the others)
 * Falling blocks can now stack up to 6 blocks (instead of 8)
 * Overtime heal now decays 5x slower. It's now configurable